+++ packages/w3c-context/src/lib/index.ts
@@ -1,39 +1,59 @@
-import { Resolver } from 'did-resolver';
+import { queryDidDocument } from '@trustvc/w3c-issuer';
 import { DocumentLoader, DocumentLoaderObject } from './types';
 // @ts-ignore: No types available for jsonld-signatures
 import jsonldSignatures from 'jsonld-signatures';
-import { getResolver as webGetResolver } from 'web-did-resolver';
 import attachmentsContext from '../context/attachments-context.json';
 import bbsV1 from '../context/bbs-v1.json';
 import bolContext from '../context/bill-of-lading.json';
+import bolcContext from '../context/bill-of-lading-carrier.json';
+import cooContext from '../context/coo.json';
 import credentialsV1 from '../context/credentials-v1.json';
 import credentialsV2 from '../context/credentials-v2.json';
+import dataIntegrityV2 from '../context/data-integrity-v2.json';
 import didV1 from '../context/did-v1.json';
 import invoiceContext from '../context/invoice.json';
 import jwsV1 from '../context/jws-2020-v1.json';
+import multikeyV1 from '../context/multikey-v1.json';
+import promissoryNoteContext from '../context/promissory-note.json';
+import qrCodeContext from '../context/qrcode-context.json';
 import renderContext from '../context/render-method-context.json';
+import statusList2021V1 from '../context/status-list-2021-v1.json';
 import trContext from '../context/transferable-records-context.json';
+import warehouseReceiptContext from '../context/warehouse-receipt.json';
 import { Document } from './types';
 
+export const DATA_INTEGRITY_V2_URL = 'https://w3id.org/security/data-integrity/v2';
 export const DID_V1_URL = 'https://www.w3.org/ns/did/v1';
 export const VC_V1_URL = 'https://www.w3.org/2018/credentials/v1';
 export const VC_V2_URL = 'https://www.w3.org/ns/credentials/v2';
+export const BBS_V1_URL = 'https://w3id.org/security/bbs/v1';
+export const BLS12381_2020_V1_URL = 'https://w3id.org/security/suites/bls12381-2020/v1';
+export const JWS_V1_URL = 'https://w3id.org/security/suites/jws-2020/v1';
+export const MULTIKEY_V1_URL = 'https://w3id.org/security/multikey/v1';
+export const STATUS_LIST_2021_CREDENTIAL_URL = 'https://w3id.org/vc/status-list/2021/v1';
+
 export const TR_CONTEXT_URL = 'https://trustvc.io/context/transferable-records-context.json';
 export const RENDER_CONTEXT_URL = 'https://trustvc.io/context/render-method-context.json';
 export const ATTACHMENTS_CONTEXT_URL = 'https://trustvc.io/context/attachments-context.json';
-export const BOL_CONTEXT_URL = 'https://trustvc.io/context/bill-of-lading.json';
-export const INVOICE_CONTEXT_URL = 'https://trustvc.io/context/invoice.json';
+export const QRCODE_CONTEXT_URL = 'https://trustvc.io/context/qrcode-context.json';
 
-export const BBS_V1_URL = 'https://w3id.org/security/bbs/v1';
-export const JWS_V1_URL = 'https://w3id.org/security/suites/jws-2020/v1';
-export const STATUS_LIST_2021_CREDENTIAL_URL = 'https://w3id.org/vc/status-list/2021/v1';
+export const BOL_CONTEXT_URL = 'https://trustvc.io/context/bill-of-lading.json';
+export const BOLC_CONTEXT_URL = 'https://trustvc.io/context/bill-of-lading-carrier.json';
+export const COO_CONTEXT_URL = 'https://trustvc.io/context/coo.json';
+export const INVOICE_CONTEXT_URL = 'https://trustvc.io/context/invoice.json';
+export const PROMISSORY_NOTE_CONTEXT_URL = 'https://trustvc.io/context/promissory-note.json';
+export const WAREHOUSE_RECEIPT_CONTEXT_URL = 'https://trustvc.io/context/warehouse-receipt.json';
 
 export const contexts: { [key: string]: Document } = {
+  [DATA_INTEGRITY_V2_URL]: dataIntegrityV2,
   [DID_V1_URL]: didV1,
   [VC_V1_URL]: credentialsV1,
   [VC_V2_URL]: credentialsV2,
   [BBS_V1_URL]: bbsV1,
+  [BLS12381_2020_V1_URL]: bbsV1,
   [JWS_V1_URL]: jwsV1,
+  [MULTIKEY_V1_URL]: multikeyV1,
+  [STATUS_LIST_2021_CREDENTIAL_URL]: statusList2021V1,
 };
 
 export const trContexts: { [key: string]: Document } = {
@@ -48,12 +68,17 @@ export const attachmentsContexts: { [key: string]: Document } = {
   [ATTACHMENTS_CONTEXT_URL]: attachmentsContext,
 };
 
-export const bolContexts: { [key: string]: Document } = {
-  [BOL_CONTEXT_URL]: bolContext,
+export const qrCodeContexts: { [key: string]: Document } = {
+  [QRCODE_CONTEXT_URL]: qrCodeContext,
 };
 
-export const invoiceContexts: { [key: string]: Document } = {
+export const templateContexts: { [key: string]: Document } = {
+  [BOL_CONTEXT_URL]: bolContext,
+  [BOLC_CONTEXT_URL]: bolcContext,
+  [COO_CONTEXT_URL]: cooContext,
   [INVOICE_CONTEXT_URL]: invoiceContext,
+  [PROMISSORY_NOTE_CONTEXT_URL]: promissoryNoteContext,
+  [WAREHOUSE_RECEIPT_CONTEXT_URL]: warehouseReceiptContext,
 };
 
 export const CredentialContextVersion = {
@@ -80,8 +105,8 @@ export async function getDocumentLoader(
     trContexts,
     renderContexts,
     attachmentsContexts,
-    bolContexts,
-    invoiceContexts,
+    qrCodeContexts,
+    templateContexts,
     additionalContexts,
   ].forEach((context) => {
     if (!context) return;
@@ -95,14 +120,10 @@ export async function getDocumentLoader(
   });
 
   const resolveDid = async (did: string) => {
-    const resolver = new Resolver({
-      ...webGetResolver(),
-    });
-    const doc = await resolver.resolve(did);
-
+    const { wellKnownDid } = await queryDidDocument({ did });
     const result: DocumentLoaderObject = {
       contextUrl: null,
-      document: doc.didDocument,
+      document: wellKnownDid,
       documentUrl: did,
     };