+++ packages/w3c-vc/src/lib/helper/index.ts
@@ -7,10 +7,11 @@ import {
   TransferableRecordsCredentialStatus,
 } from '@trustvc/w3c-credential-status';
 import { BBSPrivateKeyPair, PrivateKeyPair, VerificationType } from '@trustvc/w3c-issuer';
+import { createHash } from 'crypto';
 // @ts-ignore: No types available for jsonld
 import * as jsonld from 'jsonld';
 import { v7 as uuidv7 } from 'uuid';
-import { assertCredentialStatus } from '../sign/credentialStatus';
+import { assertCredentialStatuses } from '../sign/credentialStatus';
 import {
   CredentialStatus,
   CredentialSubject,
@@ -19,7 +20,6 @@ import {
   RawVerifiableCredential,
   VerifiableCredential,
 } from '../types';
-import { createHash } from 'crypto';
 
 /**
  * Validates a key pair object to ensure it contains the required properties.
@@ -180,7 +180,8 @@ export function _checkCredential<T extends VerifiableCredential>(
     assertDateString({ credential, prop: 'expirationDate' });
     if (mode === 'verify') {
       if (now > new Date(credential.expirationDate)) {
-        throw new Error('Credential has expired.');
+        console.warn('Credential has expired.');
+        // throw new Error('Credential has expired.');
       }
     }
   }
@@ -222,9 +223,7 @@ export function _checkCredential<T extends VerifiableCredential>(
   }
 
   // Validate credentialStatus field if present
-  jsonld.getValues(credential, 'credentialStatus').forEach((cs: CredentialStatus) => {
-    assertCredentialStatus(cs, mode);
-  });
+  assertCredentialStatuses(credential, mode);
 
   // Validate that certain fields, if present, are objects with a type property
   for (const prop of mustHaveType) {