+++ scripts/copy-files.sh
@@ -13,12 +13,17 @@ rm -rf ./public
 mkdir -p "$DEST_DIR"
 
 # List of files to copy
-FILES=("attachments-context.json" "bill-of-lading.json" "invoice.json" "render-method-context.json" "transferable-records-context.json")
+FILES=("attachments-context.json" "bill-of-lading.json" "bill-of-lading-carrier.json" "coo.json" "invoice.json" "promissory-note.json" "qrcode-context.json" "render-method-context.json" "transferable-records-context.json" "warehouse-receipt.json")
 
 # Copy each file
 for FILE in "${FILES[@]}"; do
   cp "$SOURCE_DIR/$FILE" "$DEST_DIR/"
 done
+
+echo "[[headers]]
+  for = \"/*\"
+  [headers.values]
+    Access-Control-Allow-Origin = \"*\"" > ./public/netlify.toml
 
 # Optional: output a message when done
 echo "Files have been copied to $DEST_DIR"